pyproject.toml
[project]
name = "anthropic-effective-agents"
version = "0.1.0"
description = "Exploration of AI Effective Agents based on Anthropic's blog post"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
]
